Initial Commit
This file implements a sliding 15-puzzle.  This game is played on a 4x4
grid of tile (array of integers) There are 15 (1-15) integers and 1
blank space. The goal of the game is to slide the numbers around the
blank space to the blank space, trying to make the numbers on the tiles
move to when they are in numerical order.